requirements.txt
streamlit
transformers
requests
Pillow
